requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyMySQL==1.1.0
marshmallow-sqlalchemy==0.29.0
cryptography==41.0.7
SQLAlchemy==2.0.23
flask-restx==1.3.0
pydantic==2.6.1
email-validator==2.1.0
python-dateutil==2.8.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
gunicorn==21.2.0
Flask-CORS==4.0.0
waitress==2.1.2